update editor join session
